requirements.txt
aniso8601==9.0.1
attrs==23.1.0
blinker==1.6.2
click==8.1.3
colorama==0.4.6
Flask==2.3.2
Flask-Cors==3.0.10
flask-restx==1.1.0
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
mysql-connector==2.2.9
mysql-connector-python==8.0.33
mysqlclient==2.1.1
protobuf==3.20.3
pyrsistent==0.19.3
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
SQLAlchemy==2.0.15
typing_extensions==4.6.2
Werkzeug==2.3.4
zipp==3.15.0
